updates on ksql queries

+++ ksql/scripts/views_page.sql
@@ -30,7 +30,7 @@ CREATE TABLE t_views_page_win
 -- Table to stream
 CREATE STREAM st_views_page_win (
     document_id string,
-    window_end int,
+    window_end bigint,
     count int)
 	with (
 	    KAFKA_TOPIC='t_views_page_win',
